+++ app/View/Components/Widget/Detail/Comments.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace App\View\Components\Widget\Detail;
+
+use Closure;
+use Illuminate\Contracts\View\View;
+use Illuminate\View\Component;
+
+class Comments extends Component
+{
+    /**
+     * @var string Widget name/template name
+     */
+    public $name;
+
+    public $data;
+
+    /**
+     * Create a new component instance.
+     *
+     * @return void
+     */
+    public function __construct(string $name, ?array $data)
+    {
+        $this->name = $name;
+        $this->data = $data;
+    }
+
+    /**
+     * Get the view / contents that represent the component.
+     */
+    public function render(): View|Closure|string
+    {
+        return view('components.widget.detail.comments', $this->data);
+    }
+}