+++ resources/views/errors/426.blade.php
@@ -1,16 +1,12 @@
 @extends('layouts.blank')
 
 @push('stylesheets')
-    <link rel="stylesheet" href="{{ asset('assets/css/bootstrap-markdown.min.css') }}" />
     <link rel="stylesheet" href="{{ asset('assets/css/bootstrap-tagsinput.css') }}" />
 @endpush
 
 @push('scripts')
-    <script src="{{ asset('assets/js/bootstrap-markdown.js') }}"></script>
     <script src="{{ asset('assets/js/bootstrap-tagsinput.min.js') }}"></script>
     <script src="{{ asset('assets/js/typeahead.bundle.min.js') }}"></script>
-    <script src="{{ asset('assets/js/marked.min.js') }}"></script>
-    <script src="{{ asset('assets/js/munkireport.comment.js') }}"></script>
 @endpush
 
 @section('content')